+++ .gitignore
@@ -1,4 +1,3 @@
 /.idea
 /node_modules
-*.lock
 *.log
